Move error type into separate file
